editing and deleting with authorization(395)

+++ rest-ongoing/back-end/controllers/feed.js
@@ -3,6 +3,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -44,24 +45,32 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   // Create post in db
   const post = new Post({
     title: title, 
     content: content,
-    creator: {
-      name: 'Srikanth'  
-    },
+    creator: req.userId,
     imageUrl: imageUrl
   });
 
   post.save()
   .then(result => {
-    // console.log(result);
+    return User.findById(req.userId);
+  })
+  .then(user => {
+    creator = user;
+    user.posts.push(post);
+    return user.save();
+  })
+  .then(result => {
     res.status(201).json({
       message: 'Post created successfully!',
-      post: result
+      post: post,
+      creator: {_id: creator._id, name: creator.name }
     });
-  }).catch(error => {
+  })
+  .catch(error => {
     if (!error.statusCode) {
        error.statusCode = 500;
     }
@@ -118,6 +127,13 @@ exports.updatePost = (req,res,next) => {
         error.statusCode = 404;
         throw error;
       }
+        //Checking Authorization
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not Authorized');
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -146,6 +162,13 @@ exports.deletePost = (req,res,next) => {
         error.statusCode = 404;
         throw error;
       }
+      // Checking Authorization
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not Authorized');
+        error.statusCode = 403;
+        throw error;
+      }
+
       //Check loggedin user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId)
